requirements.txt
discord~=2.3.2
redbot~=2.0.12
openai~=1.12.0
python-dotenv~=1.0.1